+++ src/index.ts
@@ -1,3 +1,3 @@
-export { typeFlag as default } from './type-flag';
+export { typeFlag } from './type-flag';
 export { getFlag } from './get-flag';
 export type { TypeFlag, Flags } from './types';
